TieFake/requirements.txt
fvcore==0.1.5.post20220414
iopath==0.1.9
joblib==1.1.0
matplotlib==3.5.1
matplotlib==3.5.2
nltk==3.6.5
nltk==3.7
numpy==1.21.4
numpy==1.21.5
numpyencoder==0.3.0
pandas==1.4.1
pandas==1.4.2
Pillow==8.4.0
Pillow==10.0.0
Requests==2.31.0
rfconv==0.0.2b20210406
scikit_learn==1.0.1
torchvision==0.10.0
tqdm==4.62.3
transformers==4.15.0
